+++ client/rpc/src/test/kotlin/net/corda/client/rpc/AbstractRPCTest.kt
@@ -4,7 +4,7 @@ import net.corda.core.internal.concurrent.flatMap
 import net.corda.core.internal.concurrent.map
 import net.corda.core.messaging.RPCOps
 import net.corda.core.utilities.seconds
-import net.corda.node.services.messaging.RPCServerConfiguration
+import net.corda.node.services.rpc.RPCServerConfiguration
 import net.corda.testing.core.SerializationEnvironmentRule
 import net.corda.testing.node.User
 import net.corda.testing.node.internal.RPCDriverDSL
