+++ node/src/main/kotlin/net/corda/node/internal/rpc/proxies/AuthenticatedRpcOpsProxy.kt
@@ -4,8 +4,8 @@ import net.corda.client.rpc.PermissionException
 import net.corda.core.flows.FlowLogic
 import net.corda.core.messaging.CordaRPCOps
 import net.corda.node.internal.InvocationHandlerTemplate
-import net.corda.node.services.messaging.RpcAuthContext
-import net.corda.node.services.messaging.rpcContext
+import net.corda.node.services.rpc.RpcAuthContext
+import net.corda.node.services.rpc.rpcContext
 import java.lang.reflect.Method
 import java.lang.reflect.Proxy
