+++ node/src/test/kotlin/net/corda/node/internal/serialization/RpcServerObservableSerializerTests.kt
@@ -7,7 +7,7 @@ import net.corda.core.context.Trace
 import net.corda.node.internal.serialization.testutils.TestObservableContext
 import net.corda.node.internal.serialization.testutils.serializationContext
 import net.corda.node.serialization.amqp.RpcServerObservableSerializer
-import net.corda.node.services.messaging.ObservableSubscription
+import net.corda.node.services.rpc.ObservableSubscription
 import net.corda.serialization.internal.AllWhitelist
 import net.corda.serialization.internal.amqp.SerializationOutput
 import net.corda.serialization.internal.amqp.SerializerFactoryBuilder
