+++ node/src/test/kotlin/net/corda/node/internal/serialization/testutils/TestObservableContext.kt
@@ -2,8 +2,8 @@ package net.corda.node.internal.serialization.testutils
 
 import com.github.benmanes.caffeine.cache.Cache
 import net.corda.core.context.Trace
-import net.corda.node.services.messaging.ObservableContextInterface
-import net.corda.node.services.messaging.ObservableSubscription
+import net.corda.node.services.rpc.ObservableContextInterface
+import net.corda.node.services.rpc.ObservableSubscription
 import net.corda.nodeapi.RPCApi
 import org.apache.activemq.artemis.api.core.SimpleString
 import java.util.concurrent.ConcurrentHashMap
